fix: apply method-based transformations and extract shared doc URL module
- Fix referDoc.ts old-guide branch: apply transformDocPathForMethod() before transformOperationToGuidePage() for multi-method endpoints
- Fix completionProvider.ts buildDocUrl: same fix for old guide URLs
- Fix es-editor Ctrl+D handler: detect isOpenSearch for correct doc engine type
- Extract shared doc URL building logic from both files into src/common/monaco/docUrl.ts
- Tests: add 9 test cases covering multi-method old-guide URL generation
- Verify: 24/24 doc URLs return HTTP 200 across all ES versions, OpenSearch, and all endpoint patterns
